[Refactor] Pull out KHeap into CKHeap.h
Pull out the implementation for binary heap into its own templated h
file.
